requirements.txt
aniso8601==9.0.1
Authlib==1.2.0
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
colorama==0.4.6
Flask==3.0.1
Flask-RESTful==0.3.10
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
loguru==0.7.2
MarkupSafe==2.1.4
pycryptodome==3.18.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
wheel==0.41.2
win32-setctime==1.1.0
spark_ai_python
